+++ ERM/src/test/NameStorageTest.java
@@ -63,7 +63,7 @@ public class NameStorageTest {
 
 		// CREATE NAME REGISTRATION
 		Transaction nameRegistration = new RegisterNameTransaction(sender,
-				name, 0, timestamp,
+				name, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		//nameRegistration.sign(sender);
 		
@@ -92,7 +92,7 @@ public class NameStorageTest {
 
 		// ADDING KEY COMPLETE WITH YES
 		ArbitraryTransactionV3 arbitraryTransaction = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 		
@@ -119,7 +119,7 @@ public class NameStorageTest {
 
 			// ADDING KEY COMPLETE WITH YES
 			arbitraryTransaction = new ArbitraryTransactionV3(
-					badSender, null, 10, data, 0, timestamp,
+					badSender, null, 10, data, (byte)0, timestamp,
 					badSender.getLastReference(databaseSet));
 			
 			
@@ -154,7 +154,7 @@ public class NameStorageTest {
 
 		// ADDING KEY COMPLETE WITH YES
 		ArbitraryTransaction arbitraryTransaction = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 		arbitraryTransaction.process(databaseSet);
@@ -174,7 +174,7 @@ public class NameStorageTest {
 		data = storageJsonObject.toString().getBytes();
 
 		arbitraryTransaction = new ArbitraryTransactionV3(sender, null, 10, data,
-				0, timestamp,
+				(byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 		arbitraryTransaction.process(databaseSet);
@@ -194,7 +194,7 @@ public class NameStorageTest {
 		data = storageJsonObject.toString().getBytes();
 
 		arbitraryTransaction = new ArbitraryTransactionV3(sender, null, 10, data,
-				0, timestamp,
+				(byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		
 		arbitraryTransaction.process(databaseSet);
@@ -224,7 +224,7 @@ public class NameStorageTest {
 
 		// ADDING Skerberus as List key
 		ArbitraryTransaction arbitraryTransaction = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 		arbitraryTransaction.process(databaseSet);
@@ -243,7 +243,7 @@ public class NameStorageTest {
 
 		// ADDING vrontis as List key
 		arbitraryTransaction = new ArbitraryTransactionV3(sender, null, 10, data,
-				0, timestamp,
+				(byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 		
@@ -264,7 +264,7 @@ public class NameStorageTest {
 
 		// removing skerberus as List key
 		arbitraryTransaction = new ArbitraryTransactionV3(sender, null, 10, data,
-				0, timestamp,
+				(byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 
@@ -290,7 +290,7 @@ public class NameStorageTest {
 
 		// removing skerberus as List key
 		arbitraryTransaction = new ArbitraryTransactionV3(sender, null, 10, data,
-				0, timestamp,
+				(byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 
@@ -316,7 +316,7 @@ public class NameStorageTest {
 
 		// removing skerberus as List key
 		arbitraryTransaction = new ArbitraryTransactionV3(sender, null, 10, data,
-				0, timestamp,
+				(byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 		arbitraryTransaction.process(databaseSet);
@@ -335,7 +335,7 @@ public class NameStorageTest {
 		data = storageJsonObject.toString().getBytes();
 
 		arbitraryTransaction = new ArbitraryTransactionV3(sender, null, 10, data,
-				0, timestamp,
+				(byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 
@@ -356,7 +356,7 @@ public class NameStorageTest {
 		data = storageJsonObject.toString().getBytes();
 
 		arbitraryTransaction = new ArbitraryTransactionV3(sender, null, 10, data,
-				0, timestamp,
+				(byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 
@@ -383,7 +383,7 @@ public class NameStorageTest {
 		byte[] data = storageJsonObject.toString().getBytes();
 
 		ArbitraryTransaction arbitraryTransaction = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 
@@ -403,7 +403,7 @@ public class NameStorageTest {
 		data = storageJsonObject.toString().getBytes();
 
 		ArbitraryTransaction arbitraryTransaction2 = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction2.sign(sender);
 
@@ -447,7 +447,7 @@ public class NameStorageTest {
 		byte[] data = storageJsonObject.toString().getBytes();
 
 		ArbitraryTransaction arbitraryTransaction = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 		arbitraryTransaction.process(databaseSet);
@@ -468,7 +468,7 @@ public class NameStorageTest {
 		data = storageJsonObject.toString().getBytes();
 
 		ArbitraryTransaction arbitraryTransaction2 = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction2.sign(sender);
 		arbitraryTransaction2.process(databaseSet);
@@ -489,7 +489,7 @@ public class NameStorageTest {
 		data = storageJsonObject.toString().getBytes();
 
 		ArbitraryTransaction arbitraryTransaction3 = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction3.sign(sender);
 		arbitraryTransaction3.process(databaseSet);
@@ -553,7 +553,7 @@ public class NameStorageTest {
 		byte[] data = storageJsonObject.toString().getBytes();
 
 		ArbitraryTransaction arbitraryTransaction = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 		arbitraryTransaction.process(databaseSet);
@@ -574,7 +574,7 @@ public class NameStorageTest {
 		data = storageJsonObject.toString().getBytes();
 
 		ArbitraryTransaction arbitraryTransaction2 = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction2.sign(sender);
 		arbitraryTransaction2.process(databaseSet);
@@ -595,7 +595,7 @@ public class NameStorageTest {
 		data = storageJsonObject.toString().getBytes();
 
 		ArbitraryTransaction arbitraryTransaction3 = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction3.sign(sender);
 		arbitraryTransaction3.process(databaseSet);
@@ -676,7 +676,7 @@ public class NameStorageTest {
 		byte[] data = storageJsonObject.toString().getBytes();
 
 		ArbitraryTransaction arbitraryTransaction = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction.sign(sender);
 		arbitraryTransaction.process(databaseSet);
@@ -697,7 +697,7 @@ public class NameStorageTest {
 		data = storageJsonObject.toString().getBytes();
 
 		ArbitraryTransaction arbitraryTransaction2 = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction2.sign(sender);
 		arbitraryTransaction2.process(databaseSet);
@@ -718,7 +718,7 @@ public class NameStorageTest {
 		data = storageJsonObject.toString().getBytes();
 
 		ArbitraryTransaction arbitraryTransaction3 = new ArbitraryTransactionV3(
-				sender, null, 10, data, 0, timestamp,
+				sender, null, 10, data, (byte)0, timestamp,
 				sender.getLastReference(databaseSet));
 		arbitraryTransaction3.sign(sender);
 		arbitraryTransaction3.process(databaseSet);
